Prepare a new C2S prototype
We place this listener in a new port, that we can later mark as
experimental. We also prepare a template for tests, that for now will
only assume a user can connect and exchange messages with a regular user
connected to the traditional c2s port.

This time the listener is supervised using ranch, which already does
everything our mongoose_tcp_listener does, plus a few more details like
using pools of supervisors instead of all the many c2s processes having
to be started and handled their termination through a single supervisor
that can become a bottleneck. Using ranch of course also means one piece
of our architecture that we don't need to maintain, but that the
industry already maintains and optimise for us. Open source!

The new mongoose_c2s_listener module implements our mongoose_listener
behaviour, that gives mongoose_listener_sup a child_spec with the ranch
definitions.