Implement a new barebones c2s process
The flow is as follows:

init casts a `connect` message, to indicate the process that it needs to
establish the tcp connection, outside of the init function that blocks
the supervisor.

Then all states are handled in a single handle_event. The event connect
in state connecting indicates to end negotiating the connection. The
event info with payload from tcp or ssl, regardless of the state, simply
parses the payload using exml_stream, and adds all parsed packets as
internal action events of the OTP's gen_statem behaviour, that is, they
will be processed immediately as the next event in the state machine,
without involved message passing.

Then, as internal events, all packets are processed, according to their
running state machine state. One specific event is expected in that
state, and an appropriate handle_* helper does the work. States can be
compound, so now instead of sticking data of a specific state into the
state-machine memory, we can make it part of the state itself, and it
will be used in the next event in that state, see `{wait_for_feature,
after auth, _}` and `{wait_for_feature, before_auth, _}`.

Code that specifies stanzas is moved to a helper module that construct
all the verbose #xmlel{} packets, to keep the c2s module smaller and
separate protocol from state machine.