Add c2s state timeouts
As connections are inherently long-lived, we'd want to avoid the
possibility of an attacker opening many connections that then only very
slowly progress, taking ports and memory from the system. If the
connection doesn't make progress fast enough towards the session
establishment, it should be dropped. This, of course, is configurable.